+++ Nostra Utils/src/source/nostrautils/file_mngt/Folder.cpp
@@ -10,15 +10,13 @@
 	#include <sys/stat.h>
 	#include <sys/types.h>
 	#include <dirent.h>
+	#include <errno.h>
+	#include <string.h>
 #endif
 
 namespace NOU::NOU_FILE_MNGT
 {
 	Folder::Folder(const Path& path) :
-		m_path(path)
-	{}
-
-	Folder::Folder(const NOU::NOU_DAT_ALG::String8 &path) :
 		m_path(path)
 	{}
 
@@ -36,19 +34,38 @@ namespace NOU::NOU_FILE_MNGT
 			DWORD lastError = GetLastError();
 
 			if (lastError == ERROR_PATH_NOT_FOUND)
-				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), NOU_CORE::ErrorCodes::PATH_NOT_FOUND, "The path to the folder was not found.");
+				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), 
+					NOU_CORE::ErrorCodes::PATH_NOT_FOUND, "The path to the folder was not found.");
 			else if (lastError == ERROR_ALREADY_EXISTS)
-				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), NOU_CORE::ErrorCodes::ALREADY_EXISTS, "The folder is already exists.");
+				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), 
+					NOU_CORE::ErrorCodes::ALREADY_EXISTS, "The folder is already exists.");
 			else
-				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), NOU_CORE::ErrorCodes::UNKNOWN_ERROR, "The folder could not be created due to unknown reasons.");
+				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), 
+					NOU_CORE::ErrorCodes::UNKNOWN_ERROR, 
+					"The folder could not be created due to unknown reasons.");
 
 			return false;
 		}
 
 #elif NOU_OS_LIBRARY == NOU_OS_LIBRARY_POSIX
 		
-       mkdir(path.getAbsolutePath().rawStr(), 0777);
+        int err = mkdir(path.getAbsolutePath().rawStr(), 0777);
 		
+		if(err == -1)
+		{
+			switch(errno)
+			{
+			case EEXIST:
+				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), 
+					NOU_CORE::ErrorCodes::ALREADY_EXISTS, "The folder is already exists.");
+				break;
+			default:
+				NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), 
+					NOU_CORE::ErrorCodes::UNKNOWN_ERROR, 
+					"The folder could not be created due to unknown reasons.");
+				break;
+			}
+		}
 #endif
        return true;
 	}
@@ -144,17 +161,16 @@ namespace NOU::NOU_FILE_MNGT
 
 		NOU_DAT_ALG::Vector<NOU::file_mngt::File> v;
 
-		NOU::NOU_DAT_ALG::String8 pattern(m_path.getAbsolutePath().rawStr());
-		pattern.append("\\*");
+		DIR* dirp = opendir(m_path.getAbsolutePath().rawStr());
 
-		DIR* dirp = opendir(pattern.rawStr());
+		if(dirp == nullptr) //abort, if opendir() failed
+			return v;
 
 		struct dirent *dstruct;
-		NOU::uint8 dir = dstruct->d_type;
 
-		while ((dstruct = readdir(dirp)) != NULL)
+		while ((dstruct = readdir(dirp)) != nullptr)
 		{
-			if(dstruct->d_type != DT_DIR)
+			if(dstruct->d_type == DT_REG)
 			{
 				File f1(m_path + dstruct->d_name);
 				v.emplaceBack(NOU::core::move(f1));
@@ -186,6 +202,36 @@ namespace NOU::NOU_FILE_MNGT
 		 {
 			 NOU_PUSH_ERROR(NOU_CORE::getErrorHandler(), NOU_CORE::ErrorCodes::UNKNOWN_ERROR, "The folder at the passed path could not be deleted!");
 		 }
+
+#endif
+	}
+
+	boolean Folder::exists() const
+	{
+		return Folder::exists(m_path);
+	}
+
+	boolean Folder::exists(const Path &path)
+	{
+#if NOU_OS_LIBRARY == NOU_OS_LIBRARY_WIN_H
+		
+		DWORD fileAttributes = GetFileAttributesA(path.getAbsolutePath().rawStr());
+
+		if (fileAttributes == INVALID_FILE_ATTRIBUTES)
+			return false;
+		else if (fileAttributes & FILE_ATTRIBUTE_DIRECTORY)
+			return true;
+		else
+			return false;
+
+#elif NOU_OS_LIBRARY == NOU_OS_LIBRARY_POSIX
+
+		struct stat sb;
+
+		if (stat(path.getAbsolutePath().rawStr(), &sb) == 0 && S_ISDIR(sb.st_mode))
+		    return true;
+		else
+			return false;
 
 #endif
 	}